Refactoried language plugins
