+++ nvim/lua/config/language/neoformat.lua
@@ -1,0 +1,2 @@
+
+vim.api.nvim_set_keymap('n', '<leader>F', ':Neoformat<CR>', { noremap = true, silent = true })
