+++ nvim/lua/config/language/nvim-dap.lua
@@ -1,0 +1,62 @@
+local vim = vim
+local dap = require('dap')
+dap.adapters.lldb = {
+  type = 'executable',
+  command = '/usr/bin/lldb-vscode', -- adjust as needed
+  name = "lldb"
+}
+
+local dap = require('dap')
+dap.configurations.cpp = {
+  {
+    name = "Launch",
+    type = "lldb",
+    request = "launch",
+    program = function()
+      return vim.fn.input('Path to executable: ', vim.fn.getcwd() .. '/', 'file')
+    end,
+    cwd = '${workspaceFolder}',
+    stopOnEntry = false,
+    args = {},
+
+    -- if you change `runInTerminal` to true, you might need to change the yama/ptrace_scope setting:
+    --
+    --    echo 0 | sudo tee /proc/sys/kernel/yama/ptrace_scope
+    --
+    -- Otherwise you might get the following error:
+    --
+    --    Error on launch: Failed to attach to the target process
+    --
+    -- But you should be aware of the implications:
+    -- https://www.kernel.org/doc/html/latest/admin-guide/LSM/Yama.html
+    runInTerminal = true,
+  },
+}
+
+
+-- If you want to use this for rust and c, add something like this:
+
+dap.configurations.c = dap.configurations.cpp
+dap.configurations.rust = dap.configurations.cpp
+
+vim.api.nvim_set_keymap("n", ",b", ":lua require'dap'.toggle_breakpoint()<CR>", { noremap = true, silent = true })
+vim.api.nvim_set_keymap("n", ",d", ":lua require'dap'.continue()<CR>", { noremap = true, silent = true })
+vim.api.nvim_set_keymap("n", ",o", ":lua require'dap'.step_over()<CR>", { noremap = true, silent = true })
+vim.api.nvim_set_keymap("n", ",i", ":lua require'dap'.step_into()<CR>", { noremap = true, silent = true })
+
+-- Nvim-dap-virtual-text
+-- virtual text deactivated (default)
+vim.g.dap_virtual_text = true
+-- show virtual text for current frame (recommended)
+vim.g.dap_virtual_text = true
+-- request variable values for all frames (experimental)
+vim.g.dap_virtual_text = 'all frames'
+
+-- DAPInstall config
+local dap_install = require("dap-install")
+
+dap_install.setup({
+  installation_path = vim.fn.stdpath("data") .. "/dapinstall/",
+  verbosely_call_debuggers = false,
+})
+
